Bug fix: drop compiler macro for dynamic-closure.
Problem 1: to preserve the specialness of bindings made special with
`(declare (special ...))`, the compiler macro would need to emit
`special` declarations for the rebindings.

Which brings us to problem 2: emitting `special` declarations causes
package lock errors in SBCL if the symbol is in a locked package: e.g.,
`*standard-output*`.

The best thing is to drop the compiler macro; `progv` has neither
problem.