Muffle warnings in inlined sequence functions
Nowadays SBCL will complain about code like:

    (append (take ...))

Because take is inlined, and not all of the branches return a list.

Not inlining take and friends would also solve the problem, but then
SBCL would have no way of knowing that (e.g.) they return a list when
called with a list.

I think muffling the warnings is the best solution for now.